Use the JWT ID of the AccessToken as look-up key
In the token en user-info endpoints we need to look-up the access-token.
The current implementation computes a unique UUID based on the value of
the signed and serialised JWT and stores this as the lookup key in the
access_tokens collection. When an access-token is presented to the token
and user-info endpoint the unique UUID is re-computed based on the JWT
and the look-up is done.

This is now simplified by using the JTI of the
de-serialised parsed JWT AccessToken. The new lookup
key of the access-token has an index in the MongoDB collection.